requirements.txt
Django~=3.2.15
djangorestframework~=3.14.0
drf-yasg~=1.21.7